+++ TapSDK/TapBootstrapSDK/TapBootstrapConstants.cs
@@ -7,5 +7,6 @@ namespace TapTap.Bootstrap
         public static string TAP_BOOTSTARP_CLZ = "com.tapsdk.bootstrap.wrapper.TapBootstrapService";
 
         public static string TAP_BOOTSTARP_IMPL = "com.tapsdk.bootstrap.wrapper.TapBootstrapServiceImpl";
+
     }
 }