+++ TapSDK/TapBootstrapSDK/TapError.cs
@@ -6,7 +6,7 @@ namespace TapTap.Bootstrap
 {
     public class TapError
     {
-        public TapErrorCode code;
+        public int code;
 
         public string errorDescription;
 
@@ -18,8 +18,7 @@ namespace TapTap.Bootstrap
             }
 
             var dic = Json.Deserialize(json) as Dictionary<string, object>;
-            var parseCode = SafeDictionary.GetValue<int>(dic, "code");
-            code = ParseCode(parseCode);
+            code = SafeDictionary.GetValue<int>(dic, "code");
             errorDescription = SafeDictionary.GetValue<string>(dic, "error_description");
         }
 
@@ -39,7 +38,7 @@ namespace TapTap.Bootstrap
 
         public TapError(int code, string errorDescription)
         {
-            this.code = ParseCode(code);
+            this.code = code;
             this.errorDescription = errorDescription;
         }
 
@@ -52,7 +51,7 @@ namespace TapTap.Bootstrap
 
         public TapError(TapErrorCode code, string errorDescription)
         {
-            this.code = code;
+            this.code = (int) code;
             this.errorDescription = errorDescription;
         }
     }